requirements.txt
requests
mcstatus
python-dotenv
tqdm
colorama
